Move existing manifold repairing functions into a new file
