+++ Polygon_mesh_processing/include/CGAL/Polygon_mesh_processing/repair_manifoldness.h
@@ -1,0 +1,291 @@
+// Copyright (c) 2020 GeometryFactory (France).
+// All rights reserved.
+//
+// This file is part of CGAL (www.cgal.org).
+//
+// $URL$
+// $Id$
+// SPDX-License-Identifier: GPL-3.0-or-later OR LicenseRef-Commercial
+//
+// Author(s)     : Mael Rouxel-Labbé
+
+#ifndef CGAL_POLYGON_MESH_PROCESSING_REPAIR_MANIFOLDNESS_H
+#define CGAL_POLYGON_MESH_PROCESSING_REPAIR_MANIFOLDNESS_H
+
+#include <CGAL/license/Polygon_mesh_processing/repair.h>
+
+#include <CGAL/Polygon_mesh_processing/internal/named_function_params.h>
+#include <CGAL/Polygon_mesh_processing/internal/named_params_helper.h>
+#include <CGAL/Polygon_mesh_processing/manifoldness.h>
+
+#include <iterator>
+#include <vector>
+
+namespace CGAL {
+namespace Polygon_mesh_processing {
+namespace internal {
+
+template <typename G>
+struct Vertex_collector
+{
+  typedef typename boost::graph_traits<G>::vertex_descriptor      vertex_descriptor;
+
+  bool has_old_vertex(const vertex_descriptor v) const { return collections.count(v) != 0; }
+  void tag_old_vertex(const vertex_descriptor v)
+  {
+    CGAL_precondition(!has_old_vertex(v));
+    collections[v];
+  }
+
+  void collect_vertices(vertex_descriptor v1, vertex_descriptor v2)
+  {
+    std::vector<vertex_descriptor>& verts = collections[v1];
+    if(verts.empty())
+      verts.push_back(v1);
+    verts.push_back(v2);
+  }
+
+  template<typename OutputIterator>
+  void dump(OutputIterator out)
+  {
+    typedef std::pair<const vertex_descriptor, std::vector<vertex_descriptor> > Pair_type;
+    for(const Pair_type& p : collections)
+      *out++ = p.second;
+  }
+
+  void dump(Emptyset_iterator) { }
+
+  std::map<vertex_descriptor, std::vector<vertex_descriptor> > collections;
+};
+
+template <typename PolygonMesh, typename VPM, typename ConstraintMap>
+typename boost::graph_traits<PolygonMesh>::vertex_descriptor
+create_new_vertex_for_sector(typename boost::graph_traits<PolygonMesh>::halfedge_descriptor sector_begin_h,
+                             typename boost::graph_traits<PolygonMesh>::halfedge_descriptor sector_last_h,
+                             PolygonMesh& pmesh,
+                             const VPM& vpm,
+                             const ConstraintMap& cmap)
+{
+  typedef typename boost::graph_traits<PolygonMesh>::vertex_descriptor    vertex_descriptor;
+  typedef typename boost::graph_traits<PolygonMesh>::halfedge_descriptor  halfedge_descriptor;
+
+  vertex_descriptor old_vd = target(sector_begin_h, pmesh);
+  vertex_descriptor new_vd = add_vertex(pmesh);
+  put(vpm, new_vd, get(vpm, old_vd));
+
+  put(cmap, new_vd, true);
+
+  set_halfedge(new_vd, sector_begin_h, pmesh);
+  halfedge_descriptor h = sector_begin_h;
+  do
+  {
+    set_target(h, new_vd, pmesh);
+
+    if(h == sector_last_h)
+      break;
+    else
+      h = prev(opposite(h, pmesh), pmesh);
+  }
+  while(h != sector_begin_h); // for safety
+  CGAL_assertion(h != sector_begin_h);
+
+  return new_vd;
+}
+
+template <typename PolygonMesh, typename NamedParameters>
+std::size_t make_umbrella_manifold(typename boost::graph_traits<PolygonMesh>::halfedge_descriptor h,
+                                   PolygonMesh& pmesh,
+                                   internal::Vertex_collector<PolygonMesh>& dmap,
+                                   const NamedParameters& np)
+{
+  typedef typename boost::graph_traits<PolygonMesh>::vertex_descriptor    vertex_descriptor;
+  typedef typename boost::graph_traits<PolygonMesh>::halfedge_descriptor  halfedge_descriptor;
+
+  using parameters::get_parameter;
+  using parameters::choose_parameter;
+
+  typedef typename GetVertexPointMap<PolygonMesh, NamedParameters>::type VertexPointMap;
+  VertexPointMap vpm = choose_parameter(get_parameter(np, internal_np::vertex_point),
+                                        get_property_map(vertex_point, pmesh));
+
+  typedef typename internal_np::Lookup_named_param_def<internal_np::vertex_is_constrained_t,
+                                                       NamedParameters,
+                                                       Constant_property_map<vertex_descriptor, bool> // default (no constraint pmap)
+                                                       >::type                  VerticesMap;
+  VerticesMap cmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
+                                      Constant_property_map<vertex_descriptor, bool>(false));
+
+  std::size_t nb_new_vertices = 0;
+
+  vertex_descriptor old_v = target(h, pmesh);
+  put(cmap, old_v, true); // store the duplicates
+
+  // count the number of borders
+  int border_counter = 0;
+  halfedge_descriptor ih = h, done = ih, border_h = h;
+  do
+  {
+    if(is_border(ih, pmesh))
+    {
+      border_h = ih;
+      ++border_counter;
+    }
+
+    ih = prev(opposite(ih, pmesh), pmesh);
+  }
+  while(ih != done);
+
+  bool is_non_manifold_within_umbrella = (border_counter > 1);
+  if(!is_non_manifold_within_umbrella)
+  {
+    const bool first_time_meeting_v = !dmap.has_old_vertex(old_v);
+    if(first_time_meeting_v)
+    {
+      // The star is manifold, so if it is the first time we have met that vertex,
+      // there is nothing to do, we just keep the same vertex.
+      set_halfedge(old_v, h, pmesh); // to ensure halfedge(old_v, pmesh) stays valid
+      dmap.tag_old_vertex(old_v); // so that we know we have met old_v already, next time, we'll have to duplicate
+    }
+    else
+    {
+      // This is not the canonical star associated to 'v'.
+      // Create a new vertex, and move the whole star to that new vertex
+      halfedge_descriptor last_h = opposite(next(h, pmesh), pmesh);
+      vertex_descriptor new_v = create_new_vertex_for_sector(h, last_h, pmesh, vpm, cmap);
+      dmap.collect_vertices(old_v, new_v);
+      nb_new_vertices = 1;
+    }
+  }
+  // if there is more than one sector, look at each sector and split them away from the main one
+  else
+  {
+    // the first manifold sector, described by two halfedges
+    halfedge_descriptor sector_start_h = border_h;
+    CGAL_assertion(is_border(border_h, pmesh));
+
+    bool should_stop = false;
+    bool is_main_sector = true;
+    do
+    {
+      CGAL_assertion(is_border(sector_start_h, pmesh));
+
+      // collect the sector and split it away if it must be
+      halfedge_descriptor sector_last_h = sector_start_h;
+      do
+      {
+        halfedge_descriptor next_h = prev(opposite(sector_last_h, pmesh), pmesh);
+
+        if(is_border(next_h, pmesh))
+          break;
+
+        sector_last_h = next_h;
+      }
+      while(sector_last_h != sector_start_h);
+      CGAL_assertion(!is_border(sector_last_h, pmesh));
+      CGAL_assertion(sector_last_h != sector_start_h);
+
+      halfedge_descriptor next_start_h = prev(opposite(sector_last_h, pmesh), pmesh);
+
+      // there are multiple CCs incident to this particular vertex, and we should create a new vertex
+      // if it's not the first umbrella around 'old_v' or not the first sector, but only not if it's
+      // both the first umbrella and first sector.
+      bool must_create_new_vertex = (!is_main_sector || dmap.has_old_vertex(old_v));
+
+      // In any case, we must set up the next pointer correctly
+      set_next(sector_start_h, opposite(sector_last_h, pmesh), pmesh);
+
+      if(must_create_new_vertex)
+      {
+        vertex_descriptor new_v = create_new_vertex_for_sector(sector_start_h, sector_last_h, pmesh, vpm, cmap);
+        dmap.collect_vertices(old_v, new_v);
+        ++nb_new_vertices;
+      }
+      else
+      {
+        // We are in the first sector and first star, ensure that halfedge(old_v, pmesh) stays valid
+        set_halfedge(old_v, sector_start_h, pmesh);
+      }
+
+      is_main_sector = false;
+      sector_start_h = next_start_h;
+      should_stop = (sector_start_h == border_h);
+    }
+    while(!should_stop);
+  }
+
+  return nb_new_vertices;
+}
+
+} // end namespace internal
+
+/// \ingroup PMP_repairing_grp
+/// duplicates all the non-manifold vertices of the input mesh.
+///
+/// @tparam PolygonMesh a model of `HalfedgeListGraph` and `MutableHalfedgeGraph`
+/// @tparam NamedParameters a sequence of \ref pmp_namedparameters "Named Parameters"
+///
+/// @param pm the surface mesh to be repaired
+/// @param np optional \ref pmp_namedparameters "Named Parameters" described below
+///
+/// \cgalNamedParamsBegin
+///    \cgalParamBegin{vertex_point_map} the property map with the points associated to the vertices of `pmesh`.
+///       The type of this map is model of `ReadWritePropertyMap`.
+///       If this parameter is omitted, an internal property map for
+///       `CGAL::vertex_point_t` should be available in `PolygonMesh`
+///    \cgalParamEnd
+///   \cgalParamBegin{vertex_is_constrained_map} a writable property map with `vertex_descriptor`
+///     as key and `bool` as `value_type`. `put(pmap, v, true)` will be called for each duplicated
+///     vertices, as well as the original non-manifold vertex in the input mesh.
+///  \cgalParamEnd
+///   \cgalParamBegin{output_iterator} a model of `OutputIterator` with value type
+///      `std::vector<vertex_descriptor>`. The first vertex of each vector is a non-manifold vertex
+///       of the input mesh, followed by the new vertices that were created to fix this precise
+///       non-manifold configuration.
+///  \cgalParamEnd
+/// \cgalNamedParamsEnd
+///
+/// \return the number of vertices created.
+template <typename PolygonMesh, typename NamedParameters>
+std::size_t duplicate_non_manifold_vertices(PolygonMesh& pmesh,
+                                            const NamedParameters& np)
+{
+  using parameters::get_parameter;
+  using parameters::choose_parameter;
+
+  typedef boost::graph_traits<PolygonMesh>                            GT;
+  typedef typename GT::halfedge_descriptor                            halfedge_descriptor;
+
+  typedef typename internal_np::Lookup_named_param_def<internal_np::output_iterator_t,
+                                                       NamedParameters,
+                                                       Emptyset_iterator>::type Output_iterator;
+
+  Output_iterator out = choose_parameter(get_parameter(np, internal_np::output_iterator),
+                                         Emptyset_iterator());
+
+  std::vector<halfedge_descriptor> non_manifold_cones;
+  non_manifold_vertices(pmesh, std::back_inserter(non_manifold_cones));
+
+  internal::Vertex_collector<PolygonMesh> dmap;
+  std::size_t nb_new_vertices = 0;
+  if(!non_manifold_cones.empty())
+  {
+    for(halfedge_descriptor h : non_manifold_cones)
+      nb_new_vertices += internal::make_umbrella_manifold(h, pmesh, dmap, np);
+
+    dmap.dump(out);
+  }
+
+  return nb_new_vertices;
+}
+
+template <class PolygonMesh>
+std::size_t duplicate_non_manifold_vertices(PolygonMesh& pmesh)
+{
+  return duplicate_non_manifold_vertices(pmesh, parameters::all_default());
+}
+
+
+} // namespace Polygon_mesh_processing
+} // namespace CGAL
+
+#endif // CGAL_POLYGON_MESH_PROCESSING_REPAIR_MANIFOLDNESS_H
